move download button to own component. clean up the naming for tipjar, neighborhoodrecs
