+++ components/WhatToEat.tsx
@@ -1,12 +1,11 @@
 import { WHAT_TO_EAT_ID } from "../utils/constants";
+import { getStreamResponse } from "../utils/getStreamResponse";
 import Page from "./Page";
 import styles from "./day.module.css";
 import { ReactElement } from "react";
 
 export default function WhatToEat({ locationName, whatToEat }): ReactElement {
-
   const subheader: string = "Where to Stay";
-  const markup = { __html: whatToEat };
 
   return (
     <>
@@ -18,7 +17,7 @@ export default function WhatToEat({ locationName, whatToEat }): ReactElement {
         <h3>What to Eat in {locationName}</h3>
         <div
           className={styles.day_description}
-          dangerouslySetInnerHTML={markup}
+          dangerouslySetInnerHTML={{ __html: whatToEat }}
         />
       </Page>
     </>
