+++ components/WhereToStay.tsx
@@ -1,12 +1,11 @@
 import { WHERE_TO_STAY_ID } from "../utils/constants";
+import { getStreamResponse } from "../utils/getStreamResponse";
 import Page from "./Page";
 import styles from "./day.module.css";
-import { ReactElement } from "react";
+import { ReactElement, useState } from "react";
 
-export default function NeighborhoodRecommendations({ locationName, whereToStay }): ReactElement {
-
+export default function WhereToStay({ locationName, whereToStay }): ReactElement {
   const subheader: string = "Where to Stay";
-  const markup = { __html: whereToStay };
 
   return (
     <>
@@ -18,7 +17,7 @@ export default function NeighborhoodRecommendations({ locationName, whereToStay 
         <h3>Where to Stay in {locationName}</h3>
         <div
           className={styles.day_description}
-          dangerouslySetInnerHTML={markup}
+          dangerouslySetInnerHTML={{ __html: whereToStay }}
         />
       </Page>
     </>
