+++ pages/index.tsx
@@ -2,17 +2,18 @@ import Head from "next/head";
 import { useState, useEffect, useRef, useLayoutEffect, ReactElement, RefObject } from "react";
 import Day from "../components/Day";
 import DayTrips from "../components/DayTrips";
-import DownloadModal from "../components/DownloadModal";
+import TipJarModal from "../components/TipJarModal";
 import styles from "./index.module.css";
 import { getStreamResponse } from "../utils/getStreamResponse";
 import isJsonString from "../utils/isJsonString";
 import fetchImage from "../utils/fetchImage";
 import * as stub from "../utils/stubData"
-import { DAY_IDS, DEFAULT_INTERESTS, DownloadButtonStatus, INIT_TRIP_LENGTH, TRAVEL_DAY_ID } from "../utils/constants";
+import { DAY_IDS, DEFAULT_INTERESTS, INIT_TRIP_LENGTH, TRAVEL_DAY_ID } from "../utils/constants";
 import {
   Activity,
   DayTrip,
   Food,
+  LoadingState,
   Meta,
   Neighborhood,
   Photo,
@@ -21,7 +22,7 @@ import {
 } from "../utils/types";
 import Form from "../components/Form";
 import TravelDay from "../components/TravelDay";
-import NeighborhoodRecommendations from "../components/NeighborhoodRecommendations";
+import WhereToStay from "../components/WhereToStay";
 import CalendarButton from "../components/CalendarButton";
 import TravelDayButton from "../components/TravelDayButton";
 import SplitPillMenu from "../components/SplitPillMenu";
@@ -33,8 +34,6 @@ export default function Home(): ReactElement {
 
   // Modal State
   const [isOpen, setIsOpen] = useState(false);
-  const [downloadButtonState, setDownloadButtonState] = useState(DownloadButtonStatus.READY);
-  const [isDownloadButtonDisabled, setIsDownloadButtonDisabled] = useState(false);
   const [isStickyHeader, setIsStickyHeader] = useState(false);
 
   //Form State
@@ -57,7 +56,6 @@ export default function Home(): ReactElement {
   const [dayTrips, setDayTrips] = useState([] as DayTrip[]);
   const [showResult, setShowResult] = useState(false);
 
-
   /** STUB DATA */
   // const [travelTips, setTravelTips] = useState(stub.mock_travelDay);
   // const [whereToStay, setWhereToStay] = useState(stub.mock_neighborhood_recs);
@@ -68,12 +66,6 @@ export default function Home(): ReactElement {
   // const [neighborhoods, setNeighborhoods] = useState(stub.mock_neighborhoods);
   // const [foods, setFood] = useState(stub.mock_foods);
   // const [showResult, setShowResult] = useState(true);
-
-  // Loading States
-  interface LoadingState {
-    days: boolean,
-    dayTrips: boolean,
-  };
 
   const [loading, setLoading] = useState({
     days: false,
@@ -139,7 +131,7 @@ export default function Home(): ReactElement {
           neighborhood={neighborhoods[i]}
           food={foods[i]}
           index={i}
-          locationName={city}
+          city={city}
           key={i}
           retry={retryDay}
         />
@@ -148,7 +140,6 @@ export default function Home(): ReactElement {
   }
 
   async function retryDay(index: number): Promise<string> {
-
     const response = await fetch("/api/generateRetryActivity", {
       method: "POST",
       headers: {
@@ -305,7 +296,6 @@ export default function Home(): ReactElement {
     getStreamResponse(data, setWhatToEat);
   }
 
-
   async function fetchTravelDay(): Promise<string> {
     if (!cityInput) return;
     const response = await fetch("/api/generateTravelDay", {
@@ -325,7 +315,6 @@ export default function Home(): ReactElement {
     }
     getStreamResponse(data, setTravelTips);
   }
-
 
   function fetchFoods(): void {
     for (let i: number = 0; i < neighborhoods.length; i++) {
@@ -414,6 +403,8 @@ export default function Home(): ReactElement {
     if (!isJsonString(jsonStr)) {
       return;
     }
+
+
     const json: Activity = JSON.parse(jsonStr);
 
     setActivities((prevState: Activity[]): Activity[] => {
@@ -476,8 +467,6 @@ export default function Home(): ReactElement {
     });
   }
 
-
-
   function initializeItineraryStates(): void {
     const initActivities: Activity[] = Array.from({ length: tripLength }, () => ({
       name: "",
@@ -512,58 +501,7 @@ export default function Home(): ReactElement {
   /*****************
   * FORM FUNCTIONS
   ******************/
-  async function onDownload(event): Promise<string> {
-    event.preventDefault();
-    setIsDownloadButtonDisabled(true)
 
-    if (!city) {
-      return;
-    };
-
-    setDownloadButtonState(DownloadButtonStatus.IN_PROGRESS);
-    try {
-      const response = await fetch("/api/createPDF", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          city: '',
-          neighborhoods: neighborhoods,
-          activities: activities,
-          foods: foods,
-          dayTrips: dayTrips
-        })
-      });
-
-      const responseData = await response.json();
-      if (response.status !== 200) {
-        throw responseData.error || new Error(`Request failed with status ${response.status}`);
-      }
-      let arr = new Uint8Array(responseData.result.data);
-      let buffer = arr.buffer
-
-      let blob = new Blob([buffer], { type: "application/pdf" });
-      const href = URL.createObjectURL(blob);
-
-      // Creating new object of PDF file
-      // Setting various property values
-      let alink = document.createElement('a');
-      document.body.appendChild(alink); // Required for this to work in FireFox
-      alink.href = href;
-      alink.download = "Trippin_Itinerary.pdf"
-      alink.target = "_blank";
-      alink.click();
-
-      setDownloadButtonState(DownloadButtonStatus.READY);
-      setIsDownloadButtonDisabled(false)
-    } catch (error) {
-      // Consider implementing your own error handling logic here
-      console.error(error);
-      setDownloadButtonState(DownloadButtonStatus.ERROR);
-      setIsDownloadButtonDisabled(false)
-    }
-  }
 
   function onSubmit(event): void {
     if (showResult) {
@@ -581,7 +519,6 @@ export default function Home(): ReactElement {
     fetchTravelDay();
     fetchWhereToStay();
     fetchWhatToEat();
-    setIsDownloadButtonDisabled(false);
   }
 
   const handleOnChange = (position) => {
@@ -615,7 +552,7 @@ export default function Home(): ReactElement {
 
   const dayUnit = tripLength === 1 ? "day" : "days";
   const capitalizedCity = city ? city.split(' ').map((word: string) => word ? word[0].toUpperCase() + word.slice(1).toLowerCase() : '').join(' ') : "";
-
+  const itineraryData = { city, neighborhoods, activities, foods, dayTrips };
   return (
     <div>
       <Head>
@@ -677,6 +614,8 @@ export default function Home(): ReactElement {
             {city ? <h3>{tripLength} {dayUnit} in {capitalizedCity}</h3> : ""}
           </div>
 
+          {/* <Loader loading={loading} /> */}
+
           {/* Travel Day */}
           {showResult && <TravelDay locationName={city} travelTips={travelTips} />}
 
@@ -690,10 +629,11 @@ export default function Home(): ReactElement {
             getInterestsString={getInterestsString}
             setMeta={setMeta}
             setDayTrips={setDayTrips}
+            setLoading={setLoading}
             meta={meta}
           />
           {/* WHERE TO STAY */}
-          {showResult && <NeighborhoodRecommendations locationName={city} whereToStay={whereToStay} />}
+          {showResult && <WhereToStay locationName={city} whereToStay={whereToStay} />}
 
           {/* WHAT TO EAT*/}
           {showResult && <WhatToEat locationName={city} whatToEat={whatToEat} />}
@@ -703,15 +643,14 @@ export default function Home(): ReactElement {
 
           {/* WHAT TO EAT */}
           {showResult && <SplitPillMenu
-            onDownload={onDownload}
-            downloadState={downloadButtonState}
-            isDownloadButtonDisabled={isDownloadButtonDisabled}
             isButtonDisabled={!showResult}
+            isLoading={loading.dayTrips || loading.days}
+            itineraryData={itineraryData}
             onClick={handleScrollToSection}
           />}
         </div>
         <div className={styles.footer}>Trippin Created by SugarJie Studios</div>
-        {isOpen && <DownloadModal onClose={onModalCloseClick} />}
+        {isOpen && <TipJarModal onClose={onModalCloseClick} />}
       </main>
     </div>
   );
